refactor(recipient): entities, features type 위치 변경
