+++ src/features/paper-list/apis/get-recipients.ts
@@ -3,10 +3,9 @@ import { AxiosError } from "axios"
 import { RECIPIENTS_LIMIT, RECIPIENTS_OFFSET } from "@/entities/recipient"
 import { axiosInstance } from "@/shared/config"
 
-import type { RecipientsParams } from "./types"
-import type { RecipientsResponse } from "@/entities/recipient"
+import type { RecipientsResponse, RecipientsParams } from "@/entities/recipient"
 
-export async function getRecipients({ limit = RECIPIENTS_LIMIT, offset = RECIPIENTS_OFFSET }: RecipientsParams = {}) {
+export async function getRecipients({ limit = RECIPIENTS_LIMIT, offset = RECIPIENTS_OFFSET }: RecipientsParams) {
   try {
     const { data } = await axiosInstance.get<RecipientsResponse>(`recipients/`, {
       params: {
