--- src/features/paper-list/apis/types.ts
@@ -1,4 +1,0 @@
-export interface RecipientsParams {
-  limit?: number
-  offset?: number
-}
