+++ src/features/paper-list/models/use-popular-recipients.ts
@@ -4,8 +4,7 @@ import { recipientQueryKey, RECIPIENTS_LIMIT, RECIPIENTS_OFFSET } from "@/entiti
 
 import { getRecipients } from "../apis/get-recipients"
 
-import type { RecipientsParams } from "../apis/types"
-import type { Recipient } from "@/entities/recipient"
+import type { Recipient, RecipientsParams } from "@/entities/recipient"
 
 export function usePopularRecipients({ limit = RECIPIENTS_LIMIT, offset = RECIPIENTS_OFFSET }: RecipientsParams) {
   return useSuspenseQuery({
